refactor/#102: ItemPostViewHolder, ItemCommentViewHolder를 inner class에서 일반 class로 변경
